requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.5
django-cleanup==6.0.0
django-environ==0.9.0
django-filter==22.1
django-firebase==0.0.7
django-js-asset==2.0.0
django-mptt==0.13.4
django-multiselectfield==0.1.12
djangorestframework==3.13.1
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
Markdown==3.3.7
msgpack==1.0.4
oauth2client==4.1.3
Pillow==9.1.1
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.2
pytz==2022.1
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
